% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNeff.R
\name{calcNeff}
\alias{calcNeff}
\title{calcNeff}
\usage{
calcNeff(s, mvec, nvec)
}
\arguments{
\item{s}{number of mutations}

\item{mvec}{the list of AF - after}

\item{nvec}{the list of AF - before}
}
\value{
N-effective (bottleneck effect size)
}
\description{
Function to calculate bottleneck effect size for each patient
}
\examples{
Neff <- calcNeff(s, mvec, nvec)
}
