% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPr.R
\name{calcPr}
\alias{calcPr}
\title{calcPr}
\usage{
calcPr(N, n, m)
}
\arguments{
\item{N}{predicted Neff}

\item{n}{the number of mutated allele counts - after}

\item{m}{the number of mutated allele counts - before}
}
\value{
Fitness likelihood
}
\description{
Function to compute Fitness likelihood
}
\examples{
Pr <- calcPr(N,n,m)
}
