% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitnessAnalysis.R
\name{FitnessAnalysis}
\alias{FitnessAnalysis}
\title{FitnessAnalysis}
\usage{
FitnessAnalysis(CH.df, Neff_Pr_List, SigOrderOfMagnitude_threshold = 5)
}
\arguments{
\item{CH.df}{a data.frame with CH mutations with the following columns::
patient_id, gene, treatment, amino_acid_change,
depth, vaf, temporal_order, pre/post_treatment,
days_from_treatment_start_to_sample_collection )}

\item{Neff_Pr_List}{Neff_Pr_List[\link{1}]: a dataframe containing Neff results for each patient;
Neff_Pr_List[\link{2}]: a dataframe containing Pr results for each patient}

\item{SigOrderOfMagnitude_threshold}{a threshhold on the order of magnitude of the likelihood of the mutation fitness
(in comparison to the lowest likelihood in that patient)}
}
\value{
prList.df : dataframe which reports the fitness of each mutation in each patient and the dynamic (pos/neg/noSelection)
}
\description{
Calculate Fitness likelihood for each variant in each patient
}
\examples{
prList.df <- FitnessAnalysis(CH.df, Neff_Pr_List)
}
