% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitnessLikelihoodAnalysis.R
\name{FitnessLikelihoodAnalysis}
\alias{FitnessLikelihoodAnalysis}
\title{FitnessLikelihoodAnalysis}
\usage{
FitnessLikelihoodAnalysis(
  CH.df,
  Neff_Pr_List,
  SigOrderOfMagnitude_threshold = 5,
  time_threshold = 540
)
}
\arguments{
\item{CH.df}{a data.frame with CH mutations with the following columns::
patient_id (combination of number and letter), gene, treatment, amino_acid_change,
depth, vaf, temporal_order, pre/post_treatment,
days_from_treatment_start_to_sample_collection )}

\item{Neff_Pr_List}{Neff_Pr_List[\link{1}]: a dataframe containing Neff results for each patient;
Neff_Pr_List[\link{2}]: a dataframe containing Pr results for each patient}

\item{SigOrderOfMagnitude_threshold}{a threshhold on the order of magnitude of the likelihood of the mutation fitness
(in comparison to the lowest likelihood in that patient)}

\item{time_threshold}{a threshhold on the timepoints (in days)}
}
\value{
prList.df : dataframe which reports the fitness of each mutation in each patient and the dynamic (pos/neg/noSelection)
}
\description{
Calculate Fitness likelihood for each variant in each patient
}
\examples{
prList.df <- FitnessLikelihoodAnalysis(CH.df, Neff_Pr_List, SigOrderOfMagnitude_threshold = 5, time_threshold = 540 )
}
