% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAnalysisPerGene.R
\name{GrowthAnalysisPerGene}
\alias{GrowthAnalysisPerGene}
\title{GrowthAnalysisPerGene}
\usage{
GrowthAnalysisPerGene(PerVarGrowth.df, time_threshold)
}
\arguments{
\item{PerVarGrowth.df}{a data.frame which is the output of GrowthAnalysisPerVariant}

\item{time_threshold}{a threshhold on the timepoints (in days)}
}
\value{
PerGeneGrowth.df : dataframe reporting the VAF-changes-per-month for each gene per treatment and other statistics
}
\description{
Growth analysis per-gene per-treatment
}
\examples{
PerGeneGrowth.df <- GrowthAnalysisPerGene(PerVarGrowth.df, time_threshold)
}
