% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAnalysisPerTreatment.R
\name{GrowthAnalysisPerTreatment}
\alias{GrowthAnalysisPerTreatment}
\title{GrowthAnalysisPerTreatment}
\usage{
GrowthAnalysisPerTreatment(PerVarGrowth.df, time_threshold)
}
\arguments{
\item{PerVarGrowth.df}{a data.frame which is the output of GrowthAnalysisPerVariant}

\item{time_threshold}{a threshhold on the timepoints (in days)}
}
\value{
PerTreatmentGrowth.df : dataframe reporting the VAF-changes-per-month for each treatment and other statistics
}
\description{
Growth analysis per-treatment
}
\examples{
PerTreatmentGrowth.df <- GrowthAnalysisPerTreatment(PerVarGrowth.df, time_threshold)

}
