% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAnalysisPerVariant.R
\name{GrowthAnalysisPerVariant}
\alias{GrowthAnalysisPerVariant}
\title{GrowthAnalysisPerVariant}
\usage{
GrowthAnalysisPerVariant(CH.df)
}
\arguments{
\item{CH.df}{a data.frame with CH mutations with the following columns::
patient_id, gene, treatment, amino_acid_change,
depth, vaf, temporal_order, pre/post_treatment,
days_from_treatment_start_to_sample_collection )}
}
\value{
PerVarGrowth.df : dataframe reporting the VAF-changes-per-month and other statistics
}
\description{
Growth analysis per-variant per-patient for each time-point
}
\examples{
PerVarGrowth.df <- GrowthAnalysisPerVariant(CH.df)
}
