% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunction.R
\name{WriteFunction}
\alias{WriteFunction}
\title{WriteFunction}
\usage{
WriteFunction(file, path)
}
\description{
Function to write the calculated analysis in a df form in the given path
}
\examples{
WriteFunction (file, path)
}
