% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunction.R
\name{WriteFunction}
\alias{WriteFunction}
\title{WriteFunction}
\usage{
WriteFunction(input.df, path, name)
}
\arguments{
\item{path}{path of the file}

\item{name}{name of the file}

\item{file}{file to be written}
}
\value{
a xlsx file written
}
\description{
Function to write the calculated analysis in a df form in the given path
}
\examples{
WriteFunction (file, path, name)
}
